Split execution module in submodules for better readability.
