+++ bonobo/strategies/base.py
@@ -1,4 +1,4 @@
-from bonobo.execution import GraphExecutionContext
+from bonobo.execution.graph import GraphExecutionContext
 
 
 class Strategy:
