+++ setup.py
@@ -40,36 +40,40 @@ setup(
     name='bonobo',
     description='Bonobo',
     license='Apache License, Version 2.0',
-    install_requires=['colorama >=0.3,<0.4', 'psutil >=5.2,<5.3', 'requests >=2.12,<2.13', 'stevedore >=1.19,<1.20'],
+    install_requires=[
+        'colorama >=0.3,<0.4', 'psutil >=5.2,<5.3', 'requests >=2.13,<2.14',
+        'stevedore >=1.19,<1.20'
+    ],
     version=version,
     long_description=read('README.rst'),
     classifiers=read('classifiers.txt', tolines),
     packages=find_packages(exclude=['ez_setup', 'example', 'test']),
     include_package_data=True,
-    data_files=[
-        (
-            'share/jupyter/nbextensions/bonobo-jupyter', [
-                'bonobo/ext/jupyter/static/extension.js', 'bonobo/ext/jupyter/static/index.js',
-                'bonobo/ext/jupyter/static/index.js.map'
-            ]
-        )
-    ],
+    data_files=[('share/jupyter/nbextensions/bonobo-jupyter', [
+        'bonobo/ext/jupyter/static/extension.js',
+        'bonobo/ext/jupyter/static/index.js',
+        'bonobo/ext/jupyter/static/index.js.map'
+    ])],
     extras_require={
         'dev': [
-            'coverage >=4.3,<4.4', 'mock >=2.0,<2.1', 'nose >=1.3,<1.4', 'pylint >=1.6,<1.7', 'pytest >=3,<4',
-            'pytest-cov >=2.4,<2.5', 'pytest-timeout >=1.2,<1.3', 'sphinx', 'sphinx_rtd_theme', 'yapf'
+            'coverage >=4.3,<4.4', 'mock >=2.0,<2.1', 'pylint >=1,<2',
+            'pytest >=3,<4', 'pytest-cov >=2,<3', 'pytest-timeout >=1,<2',
+            'sphinx', 'sphinx_rtd_theme', 'yapf'
         ],
         'jupyter': ['jupyter >=1.0,<1.1', 'ipywidgets >=6.0.0.beta5']
     },
     entry_points={
         'bonobo.commands': [
-            'init = bonobo.commands.init:register', 'run = bonobo.commands.run:register',
+            'init = bonobo.commands.init:register',
+            'run = bonobo.commands.run:register',
             'version = bonobo.commands.version:register'
         ],
         'console_scripts': ['bonobo = bonobo.commands:entrypoint'],
-        'edgy.project.features': ['bonobo = '
-                                  'bonobo.ext.edgy.project.feature:BonoboFeature']
+        'edgy.project.features':
+        ['bonobo = '
+         'bonobo.ext.edgy.project.feature:BonoboFeature']
     },
     url='https://www.bonobo-project.org/',
-    download_url='https://github.com/python-bonobo/bonobo/tarball/{version}'.format(version=version),
-)
+    download_url=
+    'https://github.com/python-bonobo/bonobo/tarball/{version}'.format(
+        version=version), )
